Tile operation
Given a width and height (px), chop an image up into tiles of that
size.
Tiles overhanging the image are padded with black (transparent) pixels.
Tiles are pushed onto the stack in reverse row-major order, so that the
top of the stack is the top-left tile as one might expect.